+++ elfloader-tool/src/arch-riscv/boot.c
@@ -59,7 +59,7 @@ char elfloader_stack_alloc[BIT(CONFIG_KERNEL_STACK_BITS)];
 
 /* first HART will initialise these */
 void *dtb = NULL;
-uint32_t dtb_size = 0;
+size_t dtb_size = 0;
 
 static int map_kernel_window(struct image_info *kernel_info)
 {
